split the enclave file into responsibilities
